Use decimal price and handle basket updates
Change ticket price storage/validation to support decimals and tighten basket handling. Migration updated price column from integer to decimal and request validators changed from integer to numeric for both store and update requests. TicketForsaleController removed an unused Request import and no longer expects a Request object for addToBasket/removeFromBasket. Frontend CartContext.removeFromCart now waits for the API response, logs failures, and only updates local state on success. Also removed a global error-suppression snippet from main.tsx.

+++ backend/app/Http/Requests/StoreOriginalTicketsRequest.php
@@ -26,7 +26,7 @@ class StoreOriginalTicketsRequest extends FormRequest
         'section' => ['required', 'string'],
         'row' => ['required', 'integer'],
         'seatNumber' => ['required', 'integer'],
-        'price' => ['required', 'integer', 'min:0'],
+        'price' => ['required', 'numeric', 'min:0'],
         'status' => ['required', 'in:pre-release,active,cancelled,expired'],
         'ticketPdfUrl' => ['required', 'string', 'url'],
         ];
